+++ scripts/postbuild.js
@@ -12,10 +12,18 @@ const schemaYamlFiles = await glob(join(sourceDir, "**/schema.*.yml")).then(
   (files) => files.map((file) => file.replace(sourceDir, "")),
 );
 
+const errorsYamlFiles = await glob(
+  join(sourceDir, "**/utils/errors.*.yml"),
+).then((files) => files.map((file) => file.replace(sourceDir, "")));
+
 for (const cmdYamlFile of cmdYamlFiles) {
   await cp(join(sourceDir, cmdYamlFile), join(distDir, cmdYamlFile));
 }
 
 for (const schemaYamlFile of schemaYamlFiles) {
   await cp(join(sourceDir, schemaYamlFile), join(distDir, schemaYamlFile));
+}
+
+for (const errorsYamlFile of errorsYamlFiles) {
+  await cp(join(sourceDir, errorsYamlFile), join(distDir, errorsYamlFile));
 }