+++ src/commands/community/subscribe_user.cmd.ts
@@ -1,4 +1,5 @@
 import { Type } from "@sinclair/typebox";
+import { errors } from "../../utils/errors.ts";
 import { parseMeta } from "../../utils/parse_meta.ts";
 import { createCommand } from "../factory.ts";
 
@@ -32,7 +33,7 @@ export const subscribeUserCmd = createCommand(
     });
 
     if (!result.success) {
-      throw new Error(`${action} fail`);
+      throw new Error(errors.action_fail.replace("{action}", action));
     }
 
     return {