+++ src/commands/creative/request_character_or_elementum.cmd.ts
@@ -1,5 +1,6 @@
 import { Type } from "@sinclair/typebox";
 import type { CharacterAssign, ElementumAssign } from "../../apis/types.ts";
+import { errors } from "../../utils/errors.ts";
 import { parseMeta } from "../../utils/parse_meta.ts";
 import { createCommand } from "../factory.ts";
 
@@ -8,13 +9,6 @@ const meta = parseMeta(
     name: Type.String(),
     title: Type.String(),
     description: Type.String(),
-    errors: Type.Object({
-      missing_id: Type.String(),
-      not_found: Type.String(),
-      type_mismatch_character: Type.String(),
-      type_mismatch_elementum: Type.String(),
-      unknown_type: Type.String(),
-    }),
     parameters: Type.Object({
       name: Type.String(),
       uuid: Type.String(),
@@ -50,7 +44,7 @@ export const requestCharacterOrElementum = createCommand(
 
     const getTcp = async () => {
       if (!uuid && !name) {
-        throw new Error(meta.errors.missing_id);
+        throw new Error(errors.tcp_missing_id_or_name);
       }
 
       if (uuid) {
@@ -85,7 +79,7 @@ export const requestCharacterOrElementum = createCommand(
     const tcp = await getTcp();
     if (!tcp) {
       throw new Error(
-        meta.errors.not_found.replace(
+        errors.tcp_not_found.replace(
           "{identifier}",
           String((name || uuid) ?? "").trim(),
         ),
@@ -94,7 +88,7 @@ export const requestCharacterOrElementum = createCommand(
 
     if (tcp.type === "oc" || tcp.type === "official") {
       if (!targetType.includes("character")) {
-        throw new Error(meta.errors.type_mismatch_character);
+        throw new Error(errors.tcp_type_mismatch_character);
       }
 
       const assignValue = {
@@ -117,7 +111,7 @@ export const requestCharacterOrElementum = createCommand(
 
     if (tcp.type === "elementum") {
       if (!targetType.includes("elementum")) {
-        throw new Error(meta.errors.type_mismatch_elementum);
+        throw new Error(errors.tcp_type_mismatch_elementum);
       }
 
       const assignValue = {
@@ -136,7 +130,7 @@ export const requestCharacterOrElementum = createCommand(
     log.warn(`request_character_or_elementum: unknown tcp type: ${tcp.type}`);
 
     throw new Error(
-      meta.errors.unknown_type.replace("{type}", String(tcp.type)),
+      errors.tcp_unknown_type.replace("{type}", String(tcp.type)),
     );
   },
 );