+++ src/commands/premium/pay_premium_order.cmd.ts
@@ -1,6 +1,7 @@
 import { Type } from "@sinclair/typebox";
 import { parseDate } from "../../utils/date.ts";
 import { IS_GLOBAL } from "../../utils/env.ts";
+import { errors } from "../../utils/errors.ts";
 import { parseMeta } from "../../utils/parse_meta.ts";
 import { createCommand } from "../factory.ts";
 
@@ -31,16 +32,16 @@ export const payPremiumOrder = createCommand(
   },
   async ({ order_uuid, channel }, { apis }) => {
     if (!IS_GLOBAL) {
-      throw new Error("This command is not supported in the current region");
+      throw new Error(errors.not_supported_in_current_region);
     }
 
     const order = await apis.commerce.order({ order_uuid });
     if (order.status !== "UNPAID") {
-      throw new Error("Order is not unpaid");
+      throw new Error(errors.order_not_unpaid);
     }
 
     if (parseDate(order.valid_until).isBefore(Date.now())) {
-      throw new Error("Order is expired");
+      throw new Error(errors.order_expired);
     }
 
     const result = await apis.commerce.pay({